+++ src/wallet/wallets/compass/CompassController.ts
@@ -7,6 +7,7 @@ import { WalletConnectV1 } from "../../walletconnect/WalletConnectV1";
 import { WalletConnectV2 } from "../../walletconnect/WalletConnectV2";
 import { ConnectedWallet } from "../ConnectedWallet";
 import { ChainInfo, WalletController } from "../WalletController";
+import { WalletError } from "../WalletError";
 import { CompassExtension } from "./CompassExtension";
 
 export class CompassController extends WalletController {
@@ -35,9 +36,11 @@ export class CompassController extends WalletController {
     if (!ext) {
       throw new Error("Compass extension is not installed");
     }
-    await ext.enable(chains.map(({ chainId }) => chainId));
+    await WalletError.wrap(ext.enable(chains.map(({ chainId }) => chainId)));
     for (const { chainId, rpc, gasPrice } of Object.values(chains)) {
-      const { bech32Address, pubKey, isNanoLedger } = await ext.getKey(chainId);
+      const { bech32Address, pubKey, isNanoLedger } = await WalletError.wrap(
+        ext.getKey(chainId)
+      );
       const key = new Secp256k1PubKey({
         chainId,
         key: pubKey,