+++ src/wallet/wallets/leap/LeapController.ts
@@ -7,6 +7,7 @@ import { onWindowEvent } from "../../utils/window";
 import { WalletConnectV2 } from "../../walletconnect/WalletConnectV2";
 import { ConnectedWallet } from "../ConnectedWallet";
 import { ChainInfo, WalletController } from "../WalletController";
+import { WalletError } from "../WalletError";
 import { LeapExtension } from "./LeapExtension";
 import { LeapWalletConnectV2 } from "./LeapWalletConnectV2";
 
@@ -32,10 +33,14 @@ export class LeapController extends WalletController {
     chains: ChainInfo<T>[]
   ) {
     const wallets = new Map<T, ConnectedWallet>();
-    await this.wc.connect(chains.map(({ chainId }) => chainId));
+    await WalletError.wrap(
+      this.wc.connect(chains.map(({ chainId }) => chainId))
+    );
     for (let i = 0; i < chains.length; i++) {
       const { chainId, rpc, gasPrice } = chains[i];
-      const { pubkey, address } = await this.wc.getAccount(chainId);
+      const { pubkey, address } = await WalletError.wrap(
+        this.wc.getAccount(chainId)
+      );
       const key = new Secp256k1PubKey({
         chainId,
         key: base64.decode(pubkey),
@@ -63,9 +68,11 @@ export class LeapController extends WalletController {
     if (!ext) {
       throw new Error("Leap extension is not installed");
     }
-    await ext.enable(chains.map(({ chainId }) => chainId));
+    await WalletError.wrap(ext.enable(chains.map(({ chainId }) => chainId)));
     for (const { chainId, rpc, gasPrice } of Object.values(chains)) {
-      const { bech32Address, pubKey, isNanoLedger } = await ext.getKey(chainId);
+      const { bech32Address, pubKey, isNanoLedger } = await WalletError.wrap(
+        ext.getKey(chainId)
+      );
       const key = new Secp256k1PubKey({
         chainId,
         key: pubKey,